Grouping data into user and cache folders
